// src/generated/api/model/authResponse.ts
/**
 * Hello Doctor Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuthResponse { 
    id?: string | null;
    username?: string | null;
    email?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    role?: string | null;
    jwtToken?: string | null;
    refreshToken?: string | null;
    refreshTokenExpiration?: string | null;
    message?: string | null;
}